add removeMissing and failOnTimeout options
Previously, plugin behavior when enabled was to replace each
repository’s URL with the URL of its mirror in the enterprise, if, _and
only if_, an enterprise mirror existed for that repo URL. If no mirror
was found, the repo was left as-is by the plugin.

However, in some cases it may be desirable to forcibly remove any repos
which are not mirrored in the enterprise. For example, security policy
may prohibit external repositories due to supply-chain security risks.

To support enforcement of such a policy via this plugin, this changeset
adds two new extension properties:

* `failOnTimeout` (default: `false`) - when true, connection timeout to
  mirrors cause build to fail
* `removeMissing` (default: `false`) - when true, any repositories
  which cannot be replaced by an appropriate mirror are removed from the
  project configuration